post flatten column names dont need top level name
Since we are doing a demux alongside this flatten, we don't need the top
level struct's name to appear in all of its members, since the demux
ensures that all fields will come either from it (or the top like
timestamp).

This is vital b/c postgres emposes a 63 character limit of column names,
and the names of these top level metrics are quite large.