Support complex arrays and maps
The Gson library doesn't understand the kafka Struct object, which means
that it gets encoded in a funky way (e.g. an array of records has each
row encoded with the schema).

To fix this, lets transform all Struct objects into a HashMap, an object
that Gson knows how to process. This is a no-op for non-complex arrays
and maps.